+++ examples/blank.py
@@ -1,0 +1,4 @@
+from drafter import *
+
+
+start_server()